useAmountValidation
